+++ lib/spack/spack/detection/__init__.py
@@ -1,0 +1,14 @@
+# Copyright 2013-2021 Lawrence Livermore National Security, LLC and other
+# Spack Project Developers. See the top-level COPYRIGHT file for details.
+#
+# SPDX-License-Identifier: (Apache-2.0 OR MIT)
+from ._common import DetectedPackage, executable_prefix, update_configuration
+from .path import by_path, executables_in_path
+
+__all__ = [
+    'DetectedPackage',
+    'by_path',
+    'executables_in_path',
+    'executable_prefix',
+    'update_configuration'
+]
